Add tests for app, move component tests to directory
